src/spark-processor/requirements.txt
# spark-processor/requirements.txt
pyspark==3.4.1
kafka-python==2.0.2
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
loguru==0.7.0
